attacker/llm/requirements.txt
selenium==4.15.2
google-generativeai==0.8.5
Pillow==10.2.0
webdriver-manager==4.0.1 

